new: adjust DGL to ArangoDB for increased accessibility to python-arango interface (#24)
* #23: initial commit

* Update build.yml

* Update README.md

* remove: unecessary dict merge using Kwargs

The DGL HeteroNodeDataView / HeteroEdgeDataView is able to track previously inserted features, so this is not needed

* remove: setuptools_scm from setup.py

deprecated in favor of pyproject.toml usage

* Update README.md

* remove: __validate_attributes()

Uncessary noise as we already have proper docstring examples

* pragma: no cover

* fix: black

* chg: #24 (after redefining scope)

* new: CodeQL Action v2

* drop python 3.6 support

mirrors https://github.com/arangoml/networkx-adapter/pull/81/commits/5930c804b03a6f5322b930a498dd68d3c5a743c9

* Update adapter.py

* fix: increase coverage

* prep for 2.1.0 release: documentation tweaks

* bump: python-arango version

+++ adbdgl_adapter/adapter.py
@@ -78,7 +78,8 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
             to DGL, along with their associated attributes to keep.
         :type metagraph: adbdgl_adapter.typings.ArangoMetagraph
         :param query_options: Keyword arguments to specify AQL query options when
-            fetching documents from the ArangoDB instance.
+            fetching documents from the ArangoDB instance. Full parameter list:
+            https://docs.python-arango.com/en/main/specs.html#arango.aql.AQL.execute
         :type query_options: Any
         :return: A DGL Heterograph
         :rtype: dgl.heterograph.DGLHeteroGraph
@@ -102,7 +103,6 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
         }
         """
         logger.debug(f"Starting arangodb_to_dgl({name}, ...):")
-        self.__validate_attributes("graph", set(metagraph), self.METAGRAPH_ATRIBS)
 
         # Maps ArangoDB vertex IDs to DGL node IDs
         adb_map: Dict[str, Dict[str, Any]] = dict()
@@ -139,7 +139,7 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
                 from_col.add(from_node["col"])
                 to_col.add(to_node["col"])
                 if len(from_col | to_col) > 2:
-                    raise ValueError(
+                    raise ValueError(  # pragma: no cover
                         f"""Can't convert to DGL:
                             too many '_from' & '_to' collections in {e_col}
                         """
@@ -182,8 +182,9 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
         :type v_cols: Set[str]
         :param e_cols: A set of ArangoDB edge collections to import to DGL.
         :type e_cols: Set[str]
-        :param query_options: Keyword arguments to specify AQL query options
-            when fetching documents from the ArangoDB instance.
+        :param query_options: Keyword arguments to specify AQL query options when
+            fetching documents from the ArangoDB instance. Full parameter list:
+            https://docs.python-arango.com/en/main/specs.html#arango.aql.AQL.execute
         :type query_options: Any
         :return: A DGL Heterograph
         :rtype: dgl.heterograph.DGLHeteroGraph
@@ -200,8 +201,9 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
 
         :param name: The ArangoDB graph name.
         :type name: str
-        :param query_options: Keyword arguments to specify AQL query options
-            when fetching documents from the ArangoDB instance.
+        :param query_options: Keyword arguments to specify AQL query options when
+            fetching documents from the ArangoDB instance. Full parameter list:
+            https://docs.python-arango.com/en/main/specs.html#arango.aql.AQL.execute
         :type query_options: Any
         :return: A DGL Heterograph
         :rtype: dgl.heterograph.DGLHeteroGraph
@@ -213,7 +215,11 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
         return self.arangodb_collections_to_dgl(name, v_cols, e_cols, **query_options)
 
     def dgl_to_arangodb(
-        self, name: str, dgl_g: Union[DGLGraph, DGLHeteroGraph], batch_size: int = 1000
+        self,
+        name: str,
+        dgl_g: Union[DGLGraph, DGLHeteroGraph],
+        overwrite_graph: bool = False,
+        **import_options: Any,
     ) -> ADBGraph:
         """Create an ArangoDB graph from a DGL graph.
 
@@ -221,8 +227,13 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
         :type name: str
         :param dgl_g: The existing DGL graph.
         :type dgl_g: Union[dgl.DGLGraph, dgl.heterograph.DGLHeteroGraph]
-        :param batch_size: The maximum number of documents to insert at once
-        :type batch_size: int
+        :param overwrite_graph: Overwrites the graph if it already exists.
+            Does not drop associated collections.
+        :type overwrite_graph: bool
+        :param import_options: Keyword arguments to specify additional
+            parameters for ArangoDB document insertion. Full parameter list:
+            https://docs.python-arango.com/en/main/specs.html#arango.collection.Collection.import_bulk
+        :type import_options: Any
         :return: The ArangoDB Graph API wrapper.
         :rtype: arango.graph.Graph
         """
@@ -237,24 +248,31 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
             else dgl_g.canonical_etypes
         )
 
-        self.__db.delete_graph(name, ignore_missing=True)
-        adb_graph: ADBGraph = self.__db.create_graph(name, edge_definitions)
+        if overwrite_graph:
+            logger.debug("Overwrite graph flag is True. Deleting old graph.")
+            self.__db.delete_graph(name, ignore_missing=True)
+
+        if self.__db.has_graph(name):
+            adb_graph = self.__db.graph(name)
+        else:
+            adb_graph = self.__db.create_graph(name, edge_definitions)
 
         adb_v_cols = adb_graph.vertex_collections()
-        adb_e_cols = [e_d["edge_collection"] for e_d in edge_definitions]
+        adb_e_cols = [e_d["edge_collection"] for e_d in adb_graph.edge_definitions()]
 
         has_one_vcol = len(adb_v_cols) == 1
         has_one_ecol = len(adb_e_cols) == 1
         logger.debug(f"Is graph '{name}' homogenous? {has_one_vcol and has_one_ecol}")
 
         adb_documents: DefaultDict[str, List[Json]] = defaultdict(list)
+
         for v_col in adb_v_cols:
-            ntype = None if is_default else v_col
             v_col_docs = adb_documents[v_col]
+            ntype = None if is_default else v_col
             features = dgl_g.node_attr_schemes(ntype).keys()
 
-            logger.debug(f"Preparing {len(dgl_g.nodes(ntype))} '{v_col}' DGL nodes")
             node: Tensor
+            logger.debug(f"Preparing {dgl_g.number_of_nodes(ntype)} '{v_col}' nodes")
             for node in dgl_g.nodes(ntype):
                 dgl_node_id = node.item()
                 adb_vertex = {"_key": str(dgl_node_id)}
@@ -267,15 +285,15 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
                     has_one_vcol,
                 )
 
-                self.__insert_adb_docs(v_col, v_col_docs, adb_vertex, batch_size)
+                v_col_docs.append(adb_vertex)
 
         from_col: str
         to_col: str
-        from_nodes: Tensor
-        to_nodes: Tensor
+        from_n: Tensor
+        to_n: Tensor
         for e_col in adb_e_cols:
-            etype = None if is_default else e_col
             e_col_docs = adb_documents[e_col]
+            etype = None if is_default else e_col
             features = dgl_g.edge_attr_schemes(etype).keys()
 
             canonical_etype = None
@@ -285,32 +303,29 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
                 canonical_etype = dgl_g.to_canonical_etype(e_col)
                 from_col, _, to_col = canonical_etype
 
-            from_nodes, to_nodes = dgl_g.edges(etype=etype)
-            logger.debug(f"Preparing {len(from_nodes)} '{e_col}' DGL edges")
-            for dgl_edge_id, (from_node, to_node) in enumerate(
-                zip(from_nodes, to_nodes)
-            ):
+            logger.debug(f"Preparing {dgl_g.number_of_edges(etype)} '{e_col}' edges")
+            for index, (from_n, to_n) in enumerate(zip(*dgl_g.edges(etype=etype))):
                 adb_edge = {
-                    "_key": str(dgl_edge_id),
-                    "_from": f"{from_col}/{str(from_node.item())}",
-                    "_to": f"{to_col}/{str(to_node.item())}",
+                    "_key": str(index),
+                    "_from": f"{from_col}/{str(from_n.item())}",
+                    "_to": f"{to_col}/{str(to_n.item())}",
                 }
                 self.__prepare_adb_attributes(
                     dgl_g.edata,
                     features,
-                    dgl_edge_id,
+                    index,
                     adb_edge,
                     e_col,
                     has_one_ecol,
                     canonical_etype,
                 )
 
-                self.__insert_adb_docs(e_col, e_col_docs, adb_edge, batch_size)
+                e_col_docs.append(adb_edge)
 
-        for col, doc_list in adb_documents.items():  # insert remaining documents
-            if doc_list:
-                logger.debug(f"Inserting last {len(doc_list)} documents into '{col}'")
-                self.__db.collection(col).import_bulk(doc_list, on_duplicate="replace")
+        for col, doc_list in adb_documents.items():  # import documents into ArangoDB
+            logger.debug(f"Inserting {len(doc_list)} documents into '{col}'")
+            result = self.__db.collection(col).import_bulk(doc_list, **import_options)
+            logger.debug(result)
 
         logger.info(f"Created ArangoDB '{name}' Graph")
         return adb_graph
@@ -395,9 +410,7 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
         for key, col_dict in features_data.items():
             for col, array in col_dict.items():
                 logger.debug(f"Inserting {len(array)} '{key}' features into '{col}'")
-                data[key] = (
-                    tensor(array) if has_one_type else {**data[key], col: tensor(array)}
-                )
+                data[key] = tensor(array) if has_one_type else {col: tensor(array)}
 
     def __prepare_adb_attributes(
         self,
@@ -433,32 +446,6 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
             tensor = data[key] if has_one_col else data[key][canonical_etype or col]
             doc[key] = self.__cntrl._dgl_feature_to_adb_attribute(key, col, tensor[id])
 
-    def __insert_adb_docs(
-        self,
-        col: str,
-        col_docs: List[Json],
-        doc: Json,
-        batch_size: int,
-    ) -> None:
-        """Insert an ArangoDB document into a list. If the list exceeds
-        batch_size documents, insert into the ArangoDB collection.
-
-        :param col: The collection name
-        :type col: str
-        :param col_docs: The existing documents data belonging to the collection.
-        :type col_docs: List[adbdgl_adapter.typings.Json]
-        :param doc: The current document to insert.
-        :type doc: adbdgl_adapter.typings.Json
-        :param batch_size: The maximum number of documents to insert at once
-        :type batch_size: int
-        """
-        col_docs.append(doc)
-
-        if len(col_docs) >= batch_size:
-            logger.debug(f"Inserting next {batch_size} batch documents into '{col}'")
-            self.__db.collection(col).import_bulk(col_docs, on_duplicate="replace")
-            col_docs.clear()
-
     def __fetch_adb_docs(
         self, col: str, attributes: Set[str], query_options: Any
     ) -> Result[Cursor]:
@@ -484,22 +471,3 @@ class ADBDGL_Adapter(Abstract_ADBDGL_Adapter):
         """
 
         return self.__db.aql.execute(aql, **query_options)
-
-    def __validate_attributes(
-        self, type: str, attributes: Set[str], valid_attributes: Set[str]
-    ) -> None:
-        """Validates that a set of attributes includes the required valid
-        attributes.
-
-        :param type: The context of the attribute validation
-            (e.g connection attributes, graph attributes, etc).
-        :type type: str
-        :param attributes: The provided attributes, possibly invalid.
-        :type attributes: Set[str]
-        :param valid_attributes: The valid attributes.
-        :type valid_attributes: Set[str]
-        :raise ValueError: If **valid_attributes** is not a subset of **attributes**
-        """
-        if valid_attributes.issubset(attributes) is False:
-            missing_attributes = valid_attributes - attributes
-            raise ValueError(f"Missing {type} attributes: {missing_attributes}")
